+++ tools/esketch/NJS/runkit.js
@@ -1,5 +1,10 @@
 const esketch = require('esketch');
 
+console.log(esketch.componentInfo());
+console.log(esketch.wrapperInfo());
+
+console.log(esketch.returnArgument("Text from JS file in RunKit !") + '\n');
+
 // Only displayed in 'Runkit'.
 "Click on the below left little triangle(s)...";