+++ tools/esketch/ZND/eSketch.php
@@ -4,8 +4,7 @@ $ZNDq_affix="esketch";	// We are not interested in a extension;
 
 // Begin of generic part.
 ZNDq_init();
-ZNDq_register( "./lib" . $ZNDq_affix . "znd.so" );
-echo "Yo !";
+ZNDq_register( $ZNDq_affix . "znd" );
 
 class ZNDq {
 	static public function wrapperInfo()