+++ tools/jreq/jreq.java
@@ -36,7 +36,6 @@ class JREq extends JREqDecl {
 	{
   String location = "";
   String osName = System.getProperty("os.name").toLowerCase();
-  String prefix = "";
 
  	System.loadLibrary( "jreq" );
 
@@ -44,10 +43,8 @@ class JREq extends JREqDecl {
    location = "h:/bin";
   else if ( osName.contains( "mac" ) ) {
    location = "/Users/bin";
-   prefix = "lib";
   } else {
    location = "/home/csimon/bin";
-   prefix= "lib";
   }
 
   if ( System.getenv( "EPEIOS_SRC" ) == null ) {
@@ -55,7 +52,7 @@ class JREq extends JREqDecl {
   }
 
  	init( location );
-//  register( "./" + prefix + JREqDecl.affix + "jre");
+//  register( "./" + JREqDecl.affix + "jre");
   }
 }
 // End of generic part.