+++ devel/CHANGELOG.md
@@ -2,7 +2,7 @@
 
 2017-10-12 :
 - N4ALLW
-  - `Register(...)` now automagically add the correct prefix and extension for the *Java*/*Node.js*/*PHP* component/addon/extension.
+  - New parameter on `Register(...)` to indicate how the prefix and extension for the *Java*/*Node.js*/*PHP* component/addon/extension should be normalized.
 
 2017-10-10 :
 - N4ALLW :